Avoid using ReduceOp->output_args() in rfactor
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52177




I'm trying to get rid of `output_args` for reductions, because they
shouldn't be necessary; it's reducing over its reduction axis, why
does it need to know where its output is going?

Rfactor is probably the trickiest place where we use output_args, but
it looks like it's mostly just carrying around the location of the
store, so use that instead.

Differential Revision: [D26420548](https://our.internmc.facebook.com/intern/diff/D26420548/)
ghstack-source-id: 121813072