[nnc] Add dummy reference to llvm::cfg::Update<BasicBlock>
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52321

We're seeing undefined references to this function in coverage builds.
I don't even know why the toolchain is trying to look for it, because it's not
actually used in our code anywhere.

Obviously dropping in a dummy reference is a workaround more than a real
solution, but I'd like to get the coverage build back online.
ghstack-source-id: 121818432

Differential Revision: [D26467484](https://our.internmc.facebook.com/intern/diff/D26467484/)