[JIT] Allow __exit__ to have a return value
**Summary**
In Python, the boolean interpretation of the return value of `__exit__` of objects that are used as context managers with `with` statements is used to determine
whether or not to propagate exceptions thrown inside the body of the with
statement. This latter feature is not possible to add to TorchScript at
the moment, but the current requirement that `__exit__` not have any
return values can make it difficult to script a context manager whose
`__exit__` *does* have a return value.

Accordingly, this commit removes the requirement that `__exit__` must
not have any return value. TorchScript does not interpret this return
value in the same way Python does (or at all), but this should make it
easier to share context managers between eager mode and script code.

**Test Plan**
This commit adds a return value to one of the context managers used in
`TestWith`.

ghstack-source-id: b3ddb88c975e68df409714e967e122de5c14dd94
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52336